Add first draft of DialInContent refactor
